requirements.txt
# ── UI ───────────────────────────────────────────────────────
streamlit>=1.32.0
plotly>=5.18.0

# ── Data Engineering ─────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0

# ── Database ─────────────────────────────────────────────────
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4              # pgvector Python adapter (RAG)

# ── LLM / IA ─────────────────────────────────────────────────
requests>=2.31.0
sentence-transformers>=2.2.2  # RAG embeddings (MiniLM)

# ── Machine Learning ─────────────────────────────────────────
scikit-learn>=1.3.0           # TF-IDF matching, churn, scoring
# prophet>=1.1.5              # Prévisions (optionnel — pip install prophet séparément)
statsmodels>=0.14.0           # Prévisions via ARIMA si prophet absent

# ── Export / Reporting ───────────────────────────────────────
openpyxl>=3.1.2               # Export Excel

# ── Observabilité ────────────────────────────────────────────
mlflow>=2.10.0                # Tracking ML (optionnel)
# prometheus-client>=0.19.0  # Métriques Prometheus (optionnel)

# ── Utils ────────────────────────────────────────────────────
python-dotenv>=1.0.0
PyPDF2>=3.0.0

# ── Dev / CI ─────────────────────────────────────────────────
# pytest>=8.0.0               # Tests unitaires (install separément)
# pytest-cov>=4.0.0           # Coverage
# ruff>=0.3.0                  # Linter
